+++ src/engine/TileMap/TileMap.ts
@@ -18,6 +18,8 @@ import { EventEmitter, EventKey, Handler, Subscription } from '../EventEmitter';
 import { CoordPlane } from '../Math/coord-plane';
 import { DebugConfig } from '../Debug';
 import { clamp } from '../Math/util';
+import { PointerComponent } from '../Input/PointerComponent';
+import { PointerEvent } from '../Input/PointerEvent';
 
 export interface TileMapOptions {
   /**
@@ -61,18 +63,45 @@ export interface TileMapOptions {
   meshingLookBehind?: number;
 }
 
-export type TileMapEvents = EntityEvents & {
+export type TilePointerEvents = {
+  pointerup: PointerEvent;
+  pointerdown: PointerEvent;
+  pointerenter: PointerEvent;
+  pointerleave: PointerEvent;
+  pointermove: PointerEvent;
+  pointercancel: PointerEvent;
+  pointerwheel: WheelEvent;
+  pointerdragstart: PointerEvent;
+  pointerdragend: PointerEvent;
+  pointerdragenter: PointerEvent;
+  pointerdragleave: PointerEvent;
+  pointerdragmove: PointerEvent;
+}
+
+export type TileMapEvents = EntityEvents & TilePointerEvents & {
   preupdate: PreUpdateEvent<TileMap>;
   postupdate: PostUpdateEvent<TileMap>;
   predraw: PreDrawEvent;
-  postdraw: PostDrawEvent
+  postdraw: PostDrawEvent;
 }
 
 export const TileMapEvents = {
   PreUpdate: 'preupdate',
   PostUpdate: 'postupdate',
   PreDraw: 'predraw',
-  PostDraw: 'postdraw'
+  PostDraw: 'postdraw',
+  PointerUp: 'pointerup',
+  PointerDown: 'pointerdown',
+  PointerEnter: 'pointerenter',
+  PointerLeave: 'pointerleave',
+  PointerMove: 'pointermove',
+  PointerCancel: 'pointercancel',
+  Wheel: 'pointerwheel',
+  PointerDrag: 'pointerdragstart',
+  PointerDragEnd: 'pointerdragend',
+  PointerDragEnter: 'pointerdragenter',
+  PointerDragLeave: 'pointerdragleave',
+  PointerDragMove: 'pointerdragmove',
 };
 
 /**
@@ -111,6 +140,7 @@ export class TileMap extends Entity {
     }
   }
 
+  public pointer: PointerComponent;
   public transform: TransformComponent;
   private _motion: MotionComponent;
   private _graphics: GraphicsComponent;
@@ -232,6 +262,8 @@ export class TileMap extends Entity {
     );
     this.addComponent(new DebugGraphicsComponent((ctx, debugFlags) => this.debug(ctx, debugFlags), false));
     this.addComponent(new ColliderComponent());
+    this.addComponent(new PointerComponent);
+    this.pointer = this.get(PointerComponent);
     this._graphics = this.get(GraphicsComponent);
     this.transform = this.get(TransformComponent);
     this._motion = this.get(MotionComponent);
@@ -270,6 +302,8 @@ export class TileMap extends Entity {
       currentCol = [];
     }
 
+    this._setupPointerToTile();
+
     this._graphics.localBounds = new BoundingBox({
       left: 0,
       top: 0,
@@ -283,6 +317,27 @@ export class TileMap extends Entity {
     this._engine = engine;
   }
 
+  private _forwardPointerEventToTile = (eventType: string) => (evt: PointerEvent) => {
+    const tile = this.getTileByPoint(evt.worldPos);
+    if (tile) {
+      tile.events.emit(eventType, evt);
+    }
+  }
+
+  private _setupPointerToTile() {
+    this.events.on('pointerup', this._forwardPointerEventToTile('pointerup'));
+    this.events.on('pointerdown', this._forwardPointerEventToTile('pointerdown'));
+    this.events.on('pointerenter', this._forwardPointerEventToTile('pointerenter'));
+    this.events.on('pointerleave', this._forwardPointerEventToTile('pointerleave'));
+    this.events.on('pointermove', this._forwardPointerEventToTile('pointermove'));
+    this.events.on('pointercancel', this._forwardPointerEventToTile('pointercancel'));
+    this.events.on('pointerwheel', this._forwardPointerEventToTile('pointerwheel'));
+    this.events.on('pointerdragstart', this._forwardPointerEventToTile('pointerdragstart'));
+    this.events.on('pointerdragend', this._forwardPointerEventToTile('pointerdragend'));
+    this.events.on('pointerdragenter', this._forwardPointerEventToTile('pointerdragenter'));
+    this.events.on('pointerdragleave', this._forwardPointerEventToTile('pointerdragleave'));
+    this.events.on('pointerdragmove', this._forwardPointerEventToTile('pointerdragmove'));
+  }
 
   private _originalOffsets = new WeakMap<Collider, Vector>();
   private _getOrSetColliderOriginalOffset(collider: Collider): Vector {
@@ -657,11 +712,13 @@ export interface TileOptions {
  * of the sprites in the array so the last one will be drawn on top. You can
  * use transparency to create layers this way.
  */
-export class Tile extends Entity {
+export class Tile {
   private _bounds: BoundingBox;
   private _geometry: BoundingBox;
   private _pos: Vector;
   private _posDirty = false;
+
+  public events = new EventEmitter<TilePointerEvents>()
 
   /**
    * Return the world position of the top left corner of the tile
@@ -820,7 +877,6 @@ export class Tile extends Entity {
   public data = new Map<string, any>();
 
   constructor(options: TileOptions) {
-    super();
     this.x = options.x;
     this.y = options.y;
     this.map = options.map;