+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Added pointer event support to `ex.TileMap`'s and individual `ex.Tile`'s
 - Added simplified `ex.coroutine` overloads, you need not pass engine as long as you are in an Excalibur lifecycle
   ```typescript
   const result = ex.coroutine(function* () {...});